[module metadata] use u8 instead of u16 to represent known attribute kind (#6409)

+++ aptos-move/e2e-move-tests/src/tests/attributes.rs
@@ -195,6 +195,6 @@ fn test_bad_view_attribute_in_compiled_module() {
 // We need this because we cannot produce a KnownAttribute directly.
 #[derive(Serialize)]
 pub struct FakeKnownAttribute {
-    kind: u16,
+    kind: u8,
     args: Vec<String>,
 }
