+++ aptos-move/framework/src/module_metadata.rs
@@ -46,7 +46,7 @@ pub struct RuntimeModuleMetadataV1 {
 /// Enumeration of potentially known attributes
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
 pub struct KnownAttribute {
-    kind: u16,
+    kind: u8,
     args: Vec<String>,
 }
 
@@ -59,13 +59,13 @@ pub enum KnownAttributeKind {
 impl KnownAttribute {
     pub fn view_function() -> Self {
         Self {
-            kind: KnownAttributeKind::ViewFunction as u16,
+            kind: KnownAttributeKind::ViewFunction as u8,
             args: vec![],
         }
     }
 
     pub fn is_view_function(&self) -> bool {
-        self.kind == (KnownAttributeKind::ViewFunction as u16)
+        self.kind == (KnownAttributeKind::ViewFunction as u8)
     }
 }
 
@@ -118,7 +118,7 @@ pub fn get_module_metadata(module: &CompiledModule) -> Option<RuntimeModuleMetad
 #[error("Unknown attribute ({}) for key: {}", self.attribute, self.key)]
 pub struct MetadataValidationError {
     pub key: String,
-    pub attribute: u16,
+    pub attribute: u8,
 }
 
 pub fn is_valid_view_function(
@@ -135,7 +135,7 @@ pub fn is_valid_view_function(
 
     Err(MetadataValidationError {
         key: fun.to_string(),
-        attribute: KnownAttributeKind::ViewFunction as u16,
+        attribute: KnownAttributeKind::ViewFunction as u8,
     })
 }
